Refactor orders handlers
