+++ web/strings.go
@@ -8,6 +8,8 @@ const (
 	wrongOrderID      = "ID заказа указан неверно"
 	missingOrderID    = "Не указан ID заказа"
 	emptyOrders       = "Вы еще не сделали ни одного заказа"
+	errLoadOrders     = "Ошибка при загрузке заказов"
+	errLoadOrderItems = "Ошибка при загрузке содержимого заказа"
 	modelNotFound     = "Модель не найдена"
 	wrongModelID      = "ID модели указан неверно"
 )
